feat(dom): set size based on src
Sets the width and height of all images on the page based on their `src` attribute.
