+++ package.json
@@ -97,6 +97,7 @@
     "husky": "^8.0.3",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "jquery": "^3.7.1",
     "lint-staged": "^14.0.1",
     "markdown-toc": "^1.2.0",
     "prettier": "^2.8.8",
